LayoutTests/fast/dom/Geolocation/geolocation-permissions-query-expected.txt

This test checks that Permissions::query() returns the true permission state of the Geolocation API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS permissionsQueryResult is "prompt"
PASS permissionsQueryResult is "granted"
PASS permissionsQueryResult is "denied"
PASS successfullyParsed is true

TEST COMPLETE
